requirements.txt
streamlit 
Pillow
openai
llama-index
llama-index-llms-openai
nltk
pypdf
networkx
matplotlib